requirements.txt
eel
Pillow
